Update default schema name in README
Changed the default schema name in the dbt connection profile as this was set to demo_user but raw_market_customers and raw_market_competitors were expecting this to be jaffle_shop.  Changing default schema here keeps things cleaner and all objects for this demo are created in the same database.